third_party/WebKit/LayoutTests/fast/storage/storage-disallowed-in-data-url.html

<!DOCTYPE html>
<html>
<head>
</head>
<body>
    <script src="../js/resources/js-test-pre.js"></script>
    <script>
        description("This tests that storage methods throw exceptions with reasonable messages inside 'data:' URLs.");
    
        var errorsSeen = 0;

        window.jsTestIsAsync = true;
        
        window.addEventListener('message', function (e) {
            errorsSeen++;
            window.exceptionMessage = e.data;
            if (errorsSeen == 1) {
                shouldBeEqualToString('exceptionMessage', "Access to 'localStorage' is denied for this document. Storage is disabled inside 'data:' URLs.");
            } else {
                shouldBeEqualToString('exceptionMessage', "Access to 'sessionStorage' is denied for this document. Storage is disabled inside 'data:' URLs.");
                finishJSTest();
            }
        });

    </script>
    <iframe src="data:text/html,<script>try { window.localStorage; } catch(e) { window.top.postMessage(e.message, '*'); }; try { window.sessionStorage; } catch(e) { window.top.postMessage(e.message, '*'); }</script>"></iframe>
    <script src="../js/resources/js-test-post.js"></script>
</body>
</html>
