third_party/WebKit/LayoutTests/fast/storage/storage-disallowed-in-data-url-expected.txt

This tests that storage methods throw exceptions with reasonable messages inside 'data:' URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS exceptionMessage is "Access to 'localStorage' is denied for this document. Storage is disabled inside 'data:' URLs."
PASS exceptionMessage is "Access to 'sessionStorage' is denied for this document. Storage is disabled inside 'data:' URLs."
PASS successfullyParsed is true

TEST COMPLETE
